requirements.txt
requests
boto3
pyspark
streamlit
python-dotenv
pyyaml
pandas